requirements.txt
pyvjoy
pyserial